+++ gateway/src/api/latest.rs
@@ -21,7 +21,7 @@ use shuttle_common::backends::auth::{AuthPublicKey, JwtAuthenticationLayer, Scop
 use shuttle_common::backends::cache::CacheManager;
 use shuttle_common::backends::metrics::{Metrics, TraceLayer};
 use shuttle_common::claims::{Scope, EXP_MINUTES};
-use shuttle_common::constants::limits::{MAX_PROJECTS_DEFAULT, MAX_PROJECTS_EXTRA};
+use shuttle_common::limits::ClaimExt;
 use shuttle_common::models::error::axum::CustomErrorPath;
 use shuttle_common::models::error::ErrorKind;
 use shuttle_common::models::{
@@ -37,6 +37,7 @@ use tokio::sync::{Mutex, MutexGuard};
 use tower::ServiceBuilder;
 use tracing::{field, instrument, trace};
 use ttl_cache::TtlCache;
+use ulid::Ulid;
 use utoipa::openapi::security::{ApiKey, ApiKeyValue, SecurityScheme};
 use utoipa::IntoParams;
 use utoipa::{Modify, OpenApi};
@@ -207,21 +208,21 @@ async fn create_project(
     CustomErrorPath(project_name): CustomErrorPath<ProjectName>,
     AxumJson(config): AxumJson<project::Config>,
 ) -> Result<AxumJson<project::Response>, Error> {
-    let is_admin = claim.scopes.contains(&Scope::Admin);
-    let project_limit: Option<u32> = if is_admin {
-        None
-    } else if claim.scopes.contains(&Scope::ExtraProjects) {
-        Some(MAX_PROJECTS_EXTRA)
-    } else {
-        Some(MAX_PROJECTS_DEFAULT)
-    };
+    // Check that the user is within their project limits.
+    let temporary_increase = project_name.starts_with("cch23-") as u32;
+    let can_create_project = claim.can_create_project(
+        service
+            .get_project_count(&name)
+            .await?
+            .saturating_sub(temporary_increase),
+    );
 
     let project = service
         .create_project(
             project_name.clone(),
             name.clone(),
-            is_admin,
-            project_limit,
+            claim.is_admin(),
+            can_create_project,
             config.idle_minutes,
         )
         .await?;
@@ -319,6 +320,24 @@ async fn delete_project(
     req: Request<Body>,
 ) -> Result<AxumJson<String>, Error> {
     let project_name = scoped_user.scope.clone();
+    let project = state.service.find_project(&project_name).await?;
+    let project_id =
+        Ulid::from_string(&project.project_id).expect("stored project id to be a valid ULID");
+
+    // Try to startup a destroyed project first
+    if project.state.is_destroyed() {
+        let handle = state
+            .service
+            .new_task()
+            .project(project_name.clone())
+            .and_then(task::restart(project_id))
+            .send(&state.sender)
+            .await?;
+
+        // Wait for the project to be ready
+        handle.await;
+    }
+
     let service = state.service.clone();
     let sender = state.sender.clone();
 
@@ -1098,16 +1117,19 @@ pub mod tests {
     use axum::headers::Authorization;
     use axum::http::Request;
     use futures::TryFutureExt;
+    use http::Method;
     use hyper::body::to_bytes;
     use hyper::StatusCode;
     use serde_json::Value;
+    use shuttle_common::constants::limits::{MAX_PROJECTS_DEFAULT, MAX_PROJECTS_EXTRA};
+    use test_context::test_context;
     use tokio::sync::mpsc::channel;
     use tokio::sync::oneshot;
     use tower::Service;
 
     use super::*;
     use crate::service::GatewayService;
-    use crate::tests::{RequestBuilderExt, World};
+    use crate::tests::{RequestBuilderExt, TestProject, World};
 
     #[tokio::test]
     async fn api_create_get_delete_projects() -> anyhow::Result<()> {
@@ -1139,7 +1161,7 @@ pub mod tests {
                 .unwrap()
         };
 
-        let delete_project = |project: &str| {
+        let stop_project = |project: &str| {
             Request::builder()
                 .method("DELETE")
                 .uri(format!("/projects/{project}"))
@@ -1196,7 +1218,7 @@ pub mod tests {
             .unwrap();
 
         router
-            .call(delete_project("matrix").with_header(&authorization))
+            .call(stop_project("matrix").with_header(&authorization))
             .map_ok(|resp| {
                 assert_eq!(resp.status(), StatusCode::OK);
             })
@@ -1222,7 +1244,7 @@ pub mod tests {
             .unwrap();
 
         router
-            .call(delete_project("reloaded").with_header(&authorization))
+            .call(stop_project("reloaded").with_header(&authorization))
             .map_ok(|resp| {
                 assert_eq!(resp.status(), StatusCode::NOT_FOUND);
             })
@@ -1287,6 +1309,95 @@ pub mod tests {
         //     })
         //     .await
         //     .unwrap();
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn api_create_project_limits() -> anyhow::Result<()> {
+        let world = World::new().await;
+        let service = Arc::new(GatewayService::init(world.args(), world.pool(), "".into()).await);
+
+        let (sender, mut receiver) = channel::<BoxedTask>(256);
+        tokio::spawn(async move {
+            while receiver.recv().await.is_some() {
+                // do not do any work with inbound requests
+            }
+        });
+
+        let mut router = ApiBuilder::new()
+            .with_service(Arc::clone(&service))
+            .with_sender(sender)
+            .with_default_routes()
+            .with_auth_service(world.context().auth_uri)
+            .into_router();
+
+        let neo_key = world.create_user("neo");
+
+        let create_project = |project: &str| {
+            Request::builder()
+                .method("POST")
+                .uri(format!("/projects/{project}"))
+                .header("Content-Type", "application/json")
+                .body("{\"idle_minutes\": 3}".into())
+                .unwrap()
+        };
+
+        let authorization = Authorization::bearer(&neo_key).unwrap();
+
+        // Creating three projects for a basic user succeeds.
+        for i in 0..MAX_PROJECTS_DEFAULT {
+            router
+                .call(create_project(format!("matrix-{i}").as_str()).with_header(&authorization))
+                .map_ok(|resp| {
+                    assert_eq!(resp.status(), StatusCode::OK);
+                })
+                .await
+                .unwrap();
+        }
+
+        // Creating one more project hits the project limit.
+        router
+            .call(create_project("resurrections").with_header(&authorization))
+            .map_ok(|resp| {
+                assert_eq!(resp.status(), StatusCode::FORBIDDEN);
+            })
+            .await
+            .unwrap();
+
+        // Create a new admin user. We can't simply make the previous user an admin, since their token
+        // will live in the auth cache without the admin scope.
+        let trinity_key = world.create_user("trinity");
+        world.set_super_user("trinity");
+        let authorization = Authorization::bearer(&trinity_key).unwrap();
+
+        // Creating more than the basic and pro limit of projects for an admin user succeeds.
+        for i in 0..MAX_PROJECTS_EXTRA + 1 {
+            router
+                .call(create_project(format!("reloaded-{i}").as_str()).with_header(&authorization))
+                .map_ok(|resp| {
+                    assert_eq!(resp.status(), StatusCode::OK);
+                })
+                .await
+                .unwrap();
+        }
+
+        Ok(())
+    }
+
+    #[test_context(TestProject)]
+    #[tokio::test]
+    async fn api_delete_project_that_is_ready(project: &mut TestProject) -> anyhow::Result<()> {
+        project.router_call(Method::DELETE, "/delete").await;
+
+        Ok(())
+    }
+
+    #[test_context(TestProject)]
+    #[tokio::test]
+    async fn api_delete_project_that_is_destroyed(project: &mut TestProject) -> anyhow::Result<()> {
+        project.destroy_project().await;
+        project.router_call(Method::DELETE, "/delete").await;
 
         Ok(())
     }